tests/wpt/tests/html/canvas/offscreen/layers/2d.layer.beginLayer-options.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.layer.beginLayer-options</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.beginLayer-options</h1>
<p class="desc">Checks beginLayer works for different option parameter values</p>


<script>
test(t => {
  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.beginLayer(); ctx.endLayer();
  ctx.beginLayer(null); ctx.endLayer();
  ctx.beginLayer(undefined); ctx.endLayer();
  ctx.beginLayer([]); ctx.endLayer();
  ctx.beginLayer({}); ctx.endLayer();

  assert_throws_js(TypeError, function() { ctx.beginLayer(''); });
  assert_throws_js(TypeError, function() { ctx.beginLayer(0); });
  assert_throws_js(TypeError, function() { ctx.beginLayer(1); });
  assert_throws_js(TypeError, function() { ctx.beginLayer(true); });
  assert_throws_js(TypeError, function() { ctx.beginLayer(false); });

  ctx.beginLayer({filter: null}); ctx.endLayer();
  ctx.beginLayer({filter: undefined}); ctx.endLayer();
  ctx.beginLayer({filter: []}); ctx.endLayer();
  ctx.beginLayer({filter: {}}); ctx.endLayer();
  ctx.beginLayer({filter: {name: "unknown"}}); ctx.endLayer();
  ctx.beginLayer({filter: ''}); ctx.endLayer();

  // These cases don't throw TypeError since they can be casted to a
  // DOMString.
  ctx.beginLayer({filter: 0}); ctx.endLayer();
  ctx.beginLayer({filter: 1}); ctx.endLayer();
  ctx.beginLayer({filter: true}); ctx.endLayer();
  ctx.beginLayer({filter: false}); ctx.endLayer();

}, "Checks beginLayer works for different option parameter values");
</script>
